gist-agent/requirements.txt
arxiv
requests
PyPDF2
google-generativeai
python-dotenv
